상품 할인율과 평균 평점 Java와 DB에 반영 및 item_sales(판매량) 타입 number로 변경 및 item_main varchar2로 타입 변경

+++ TeamProject/src/dao/PostScriptDAO.java
@@ -110,7 +110,7 @@ public class PostScriptDAO {
 					pVo.setPost_date(rs.getString(4));
 					pVo.setPost_help(rs.getInt(5));
 					pVo.setPost_hits(rs.getInt(6));
-					pVo.setPost_stars(rs.getInt(7));
+					pVo.setPost_stars(rs.getDouble(7));
 					pVo.setPost_content(rs.getString(8));
 				}
 			} catch (Exception e) {
